remove position absolute from navbar; move LoggedInNav, LoggedOutNav closer; NavElement
