+++ src/components/core/Page/Navbar/LoggedOutNav.tsx
@@ -1,10 +1,13 @@
 import TextButton from 'components/core/Button/TextButton';
 import useAuthModal from 'hooks/useAuthModal';
+import NavElement from './NavElement';
 
 function LoggedOutNav() {
   const showAuthModal = useAuthModal();
 
-  return <TextButton onClick={showAuthModal} text="Sign In" />;
+  return <NavElement>
+    <TextButton onClick={showAuthModal} text="Sign In" />
+  </NavElement>;
 }
 
 export default LoggedOutNav;
