+++ src/components/core/Page/Navbar/NavElement.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+// Items in the navbar should be wrapped in NavElement
+const NavElement = styled.div`
+  display: flex;
+  align-items: center;
+`;
+
+export default NavElement;
